requirements.txt
hidapi
numpy
opencv-python
openrgb-python
PyQt6
pyyaml
